kites/keygen: rename gateway -> keygen
