+++ go/src/koding/kites/keygen/export_test.go
@@ -1,4 +1,4 @@
-package gateway
+package keygen
 
 import (
 	"io"
