+++ go/src/koding/kites/keygen/gateway.go
@@ -1,4 +1,4 @@
-package gateway
+package keygen
 
 import (
 	"errors"
