+++ go/src/koding/kites/keygen/s3.go
@@ -1,4 +1,4 @@
-package gateway
+package keygen
 
 import (
 	"encoding/json"
